# API-spec.yaml
swagger: '2.0'
info:
  title: CRM API
  version: v2
paths:
...
  /DelegatedTasks:
    get:
      tags:
        - DelegatedTasks
      summary: Returns list with tasks.
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: query
          name: myTasks
          type: boolean
          description: "if true, returns tasks that were created by user sending the request"
        - in: query
          name: sharedTasks
          type: boolean
          description: "if true, returns shared tasks"
        - in: query
          name: accepted
          type: boolean
          description: "if true, returns accepted tasks"
        - in: query
          name: refused
          type: boolean
          description: "if true, returns refused tasks"
        - in: query
          name: done
          type: boolean
          description: "if true, returns done tasks"
        - in: query
          name: activePage
          type: integer
          format: int32
          description: "active page of paginated list"
        - in: query
          name: pageSize
          type: integer
          format: int32
          description: "count of tasks returned per page"
      responses:
        '200':
          description: Returns the list
          schema:
            $ref: '#/definitions/Application.DelegatedTasks.CompleteTasksDataViewModel'
        '500':
          description: Server error.
    post:
      tags:
        - DelegatedTasks
      summary: Adds the task.
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Application.DelegatedTasks.DelegatedTaskViewModel'
      responses:
        '204':
          description: Adds the task.
        '400':
          description: Failed validation (fluent validation).
          schema:
            $ref: '#/definitions/Microsoft.AspNetCore.Mvc.ProblemDetails'
        '500':
          description: Problem saving changes.
...
  '/DelegatedTasks/{id}':
    get:
      tags:
        - DelegatedTasks
      summary: Returns details about the task.
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          format: uuid
          description: "task id"
      responses:
        '200':
          description: Returns details about the task.
          schema:
            $ref: '#/definitions/Application.DelegatedTasks.DelegatedTaskViewModel'
        '404':
          description: Task not found.
        '500':
          description: Server error.
    put:
      tags:
        - DelegatedTasks
      summary: Edits the task.
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: "task id"
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Application.DelegatedTasks.DelegatedTaskViewModel'
      responses:
        '204':
          description: Task edited successfully.
        '400':
          description: Failed validation (fluent validation).
          schema:
            $ref: '#/definitions/Microsoft.AspNetCore.Mvc.ProblemDetails'
        '404':
          description: Task not found.
          schema:
            $ref: '#/definitions/Microsoft.AspNetCore.Mvc.ProblemDetails'
        '500':
          description: Problem saving changes.
    delete:
      tags:
        - DelegatedTasks
      summary: Deletes the task.
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          format: uuid
          description: "task id"
      responses:
        '204':
          description: Task deleted successfully.
        '404':
          description: Task not found.
          schema:
            $ref: '#/definitions/Microsoft.AspNetCore.Mvc.ProblemDetails'
        '500':
          description: Problem saving changes.
  '/DelegatedTasks/{id}/share/{username}':
    post:
      tags:
        - DelegatedTasks
      summary: Shares the task with other user.
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: id
          required: true
          type: string
          format: uuid
          description: "task id"
        - in: path
          name: username
          required: true
          type: string
          description: "username of the user to share task with"
      responses:
        '204':
          description: Task deleted successfully.
        '400':
          description: The task is shared already.
          schema:
            $ref: '#/definitions/Microsoft.AspNetCore.Mvc.ProblemDetails'
        '404':
          description: User to share with not found.
          schema:
            $ref: '#/definitions/Microsoft.AspNetCore.Mvc.ProblemDetails'
        '500':
          description: Problem saving changes.
  '/DelegatedTasks/accept/{id}':
    post:
      tags:
        - DelegatedTasks
      summary: Accepts shared task.
      parameters:
        - in: path
          name: id
          required: true
          type: string
          format: uuid
          description: "task id"
      responses:
        '200':
          description: Success
        '204':
          description: Task accepted.
        '404':
          description: Task not found.
        '500':
          description: Problem saving changes.
  '/DelegatedTasks/refuse/{id}':
    post:
      tags:
        - DelegatedTasks
      summary: Refuses shared task.
      parameters:
        - in: path
          name: id
          required: true
          type: string
          format: uuid
          description: "task id"
      responses:
        '200':
          description: Success
        '204':
          description: Task refused.
        '404':
          description: Task not found.
        '500':
          description: Problem saving changes.
  '/DelegatedTasks/finish/{id}':
    post:
      tags:
        - DelegatedTasks
      summary: Closes shared task.
      parameters:
        - in: path
          name: id
          required: true
          type: string
          format: uuid
          description: "task id"
      responses:
        '200':
          description: Success
        '204':
          description: Task closed successfully.
        '404':
          description: Task not found.
        '500':
          description: Problem saving changes.
...
  /Operations:
    get:
      tags:
        - Operations
      summary: Returns all operations.
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: Returns all operations.
          schema:
            $ref: '#/definitions/Application.Operations.ViewModels.CompleteStatsViewModel'
        '500':
          description: Server error.
  /Operations/count:
    get:
      tags:
        - Operations
      summary: Returns count of all operations.
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: Returns operations count.
          schema:
            format: int32
            type: integer
        '500':
          description: Server error.
...
definitions:
...
  Application.DelegatedTasks.CompleteTasksDataViewModel:
    type: object
    properties:
      tasks:
        type: array
        items:
          $ref: '#/definitions/Application.DelegatedTasks.DelegatedTaskViewModel'
      tasksCount:
        format: int32
        type: integer
  Application.DelegatedTasks.DelegatedTaskViewModel:
    type: object
    properties:
      id:
        format: uuid
        type: string
      type:
        type: string
      dateStarted:
        format: date-time
        type: string
      deadline:
        format: date-time
        type: string
      notes:
        type: string
      finishedBy:
        type: string
      accepted:
        type: boolean
      refused:
        type: boolean
      pending:
        type: boolean
      done:
        type: boolean
      userTask:
        $ref: '#/definitions/Application.DelegatedTasks.UserAccessViewModel'
  Application.DelegatedTasks.UserAccessViewModel:
    type: object
    properties:
      createdByUsername:
        type: string
      createdBy:
        type: string
      sharedWithUsername:
        type: string
      sharedWith:
        type: string
...
  Application.Operations.TotalsViewModel:
    type: object
    properties:
      leadsTotal:
        format: int32
        type: integer
      opportunitiesTotal:
        format: int32
        type: integer
      quotesTotal:
        format: int32
        type: integer
      invoicesTotal:
        format: int32
        type: integer
      conversionsTotal:
        format: int32
        type: integer
      ordersTotal:
        format: int32
        type: integer
      revenueTotal:
        format: double
        type: number
      sourcesTotal:
        $ref: '#/definitions/Application.Operations.ViewModels.SourcesViewModel'
  Application.Operations.ViewModels.CollectedOperationDataViewModel:
    type: object
    properties:
      leads:
        format: int32
        type: integer
      opportunities:
        format: int32
        type: integer
      quotes:
        format: int32
        type: integer
      invoices:
        format: int32
        type: integer
      conversions:
        format: int32
        type: integer
      orders:
        format: int32
        type: integer
      revenue:
        format: double
        type: number
      sources:
        $ref: '#/definitions/Application.Operations.ViewModels.SourcesViewModel'
      dateStart:
        format: date-time
        type: string
      dateEnd:
        format: date-time
        type: string
  Application.Operations.ViewModels.CompleteStatsViewModel:
    type: object
    properties:
      thisMonthTotals:
        $ref: '#/definitions/Application.Operations.TotalsViewModel'
      oneMonthTotals:
        $ref: '#/definitions/Application.Operations.TotalsViewModel'
      sixMonthsTotals:
        $ref: '#/definitions/Application.Operations.TotalsViewModel'
      thisMonthData:
        type: array
        items:
          $ref: '#/definitions/Application.Operations.ViewModels.CollectedOperationDataViewModel'
      oneMonthData:
        type: array
        items:
          $ref: '#/definitions/Application.Operations.ViewModels.CollectedOperationDataViewModel'
      sixMonthsData:
        type: array
        items:
          $ref: '#/definitions/Application.Operations.ViewModels.CollectedOperationDataViewModel'
      oneMonthOpportunitiesByUser:
        type: array
        items:
          $ref: '#/definitions/Application.Operations.ViewModels.OpportunitiesByUserViewModel'
      sixMonthsOpportunitiesByUser:
        type: array
        items:
          $ref: '#/definitions/Application.Operations.ViewModels.OpportunitiesByUserViewModel'
  Application.Operations.ViewModels.OpportunitiesByUserViewModel:
    type: object
    properties:
      userDisplayName:
        type: string
      leadsTotal:
        format: int32
        type: integer
      opportunitiesTotal:
        format: int32
        type: integer
  Application.Operations.ViewModels.SourcesViewModel:
    type: object
    properties:
      web:
        format: int32
        type: integer
      socialMedia:
        format: int32
        type: integer
      flyers:
        format: int32
        type: integer
      commercial:
        format: int32
        type: integer
      formerClient:
        format: int32
        type: integer
...
  Microsoft.AspNetCore.Mvc.ProblemDetails:
    type: object
    properties:
      type:
        type: string
      title:
        type: string
      status:
        format: int32
        type: integer
      detail:
        type: string
      instance:
        type: string
    additionalProperties: {}